Move private modules to underscored names
