+++ perception/hashers/__init__.py
@@ -1,11 +1,11 @@
 from .hasher import ImageHasher, VideoHasher
-from .image.average import AverageHash
-from .image.dhash import DHash
-from .image.opencv import BlockMean, ColorMoment, MarrHildreth
-from .image.phash import PHash, PHashF, PHashU8
-from .image.wavelet import WaveletHash
-from .video.framewise import FramewiseHasher
-from .video.tmk import TMKL1, TMKL2
+from .image._average import AverageHash
+from .image._dhash import DHash
+from .image._opencv import BlockMean, ColorMoment, MarrHildreth
+from .image._phash import PHash, PHashF, PHashU8
+from .image._wavelet import WaveletHash
+from .video._framewise import FramewiseHasher
+from .video._tmk import TMKL1, TMKL2
 
 __all__ = [
     "ImageHasher",
@@ -25,7 +25,7 @@ __all__ = [
 ]
 
 try:
-    from .image.pdq import PDQHash as PDQHash, PDQHashF as PDQHashF
+    from .image._pdq import PDQHash as PDQHash, PDQHashF as PDQHashF
 except ImportError:
     pass
 else:
