+++ perception/hashers/image/_average.py
