+++ perception/hashers/image/_dhash.py
