+++ perception/hashers/image/_opencv.py
