+++ perception/hashers/image/_pdq.py
