+++ perception/hashers/image/_wavelet.py
