+++ perception/hashers/video/__init__.py
@@ -1,4 +1,4 @@
-from .framewise import FramewiseHasher
-from .tmk import TMKL1, TMKL2
+from ._framewise import FramewiseHasher
+from ._tmk import TMKL1, TMKL2
 
 __all__ = ["FramewiseHasher", "TMKL1", "TMKL2"]
