+++ perception/hashers/video/_framewise.py
