+++ tests/test_hashers.py
@@ -4,7 +4,7 @@ import string
 import pytest
 
 from perception import hashers, testing
-from perception.hashers.image.pdq import PDQHash
+from perception.hashers.image._pdq import PDQHash
 
 TEST_IMAGES = [os.path.join("tests", "images", f"image{n}.jpg") for n in range(1, 11)]
